+++ src/Core/Logger.php
@@ -82,8 +82,8 @@ class Logger extends AbstractLogger
         $context = [];
 
         if (!empty($result['response'])) {
-            // ensure it is a string 
-            if (is_string($result['response'])){
+            // ensure it is a string
+            if (is_string($result['response'])) {
                 $response = PhpHelper::jsonToArray($result['response']);
             } else {
                 $response = $result['response'];