+++ src/Model/Order.php
@@ -162,11 +162,7 @@ class Order extends Order_parent
             case "AMZ_AUTH_AND_CAPT_FAILED":
                 $remark = 'AmazonPay: ERROR';
                 if (!empty($data['result']['response'])) {
-                    if (is_string($data['result']['response'])) {
-                        $response = PhpHelper::jsonToArray($data['result']['response']);
-                    } else {
-                        $response = $data['result']['response'];
-                    }
+                    $response = PhpHelper::prepareResponse($data['result']['response']);
                     $remark .= ' (' . $response['reasonCode'] . ')';
                 }
                 $this->_setFieldData('oxtransstatus', 'NOT_FINISHED');
@@ -203,11 +199,7 @@ class Order extends Order_parent
             case "AMZ_AUTH_OR_CAPT_DECLINED":
                 $remark = 'AmazonPay: Auth or Capture Declined';
                 if (!empty($data['result']['response'])) {
-                    if (is_string($data['result']['response'])) {
-                        $response = PhpHelper::jsonToArray($data['result']['response']);
-                    } else {
-                        $response = $data['result']['response'];
-                    }
+                    $response = PhpHelper::prepareResponse($data['result']['response']);
                     $remark .= ' (' . $response['reasonCode'] . ')';
                 }
                 $this->_setFieldData('oxtransstatus', 'NOT_FINISHED');